Fix warnings "instance variable not initialized"
There was a problem with variables @schema and @constructor_type
when using and loading dry-struct. The place where those should be initialized -
missed. After experimenting a bit I found on which cases we use  variables
without initialization:
  - when we extending Dry::Struct with ClassInterface
  - when we inheriting any class from Dry::Struct or its successor

This all caused by the rule that class ivars are not shared with
successors. So we need to initialize them in a hook or like that.

In my solution:
  - @constructor_type is initialized by its writer in body of Dry::Struct class
  - @schema is initialized in ```extended``` and ```inherited``` hooks

Also there was a huge warning "loading in progress, circular require..."
in lib/struct/value.rb, now - seems fixed.

Closes https://github.com/dry-rb/dry-types/issues/77