Fixed warnings about use of undefined instance variables
I removed where `instance_variable_set` was used to attempt to set
`@schema` because that was being called on the class rather than 
an instance.

For @constructor_type, I reorganized the method to use the
memoization pattern to assign the constructor_type to 
the default of strict (which was the intention of the
code as I read it).

Specs pass.